modifySauce

+++ controllers/sauce.js
@@ -30,10 +30,12 @@ exports.getOneSauce = (req, res ,next) => {
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
+            //si l'utilisateur connecté n'est pas celui qui a créé la recette
             if (sauce.userId !== req.auth.userId){
                 res.status(401).json({ error : new error('Requête non autorisée !')});
             }
             else {
+                //si c'est le bon utilisateur on supprime l'image du dossier images et on supprime la recette
             const filename = sauce.imageUrl.split('/images/')[1];
             fs.unlink(`images/${filename}`, () => {
                 Sauce.deleteOne({ _id: req.params.id})
@@ -46,20 +48,32 @@ exports.deleteSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
+    //on vérifie si il y a un changement d'image ou non
     const sauceObject = req.file ?
     {
+        //si il y a changement d'image on récupère la nouvelle URL
       ...JSON.parse(req.body.sauce),
       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } 
+    // sinon on récupère les autres changements
     : { ...req.body };
+    
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
+            //si il y a eu changement d'images on supprime du fichier images l'ancienne image puis on sauvegarde les changements
+            if (sauceObject.imageUrl !== undefined ) {
             const filename = sauce.imageUrl.split('/images/')[1];
             fs.unlink(`images/${filename}`, () => {
                 Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
                     .then(() => res.status(200).json({ message: 'sauce modifiée !'}))
                 .catch(error => res.status(400).json({ error }));
-            });
+            });}
+            //sinon on sauvegare les changements
+            else{
+                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'sauce modifiée !'}))
+                .catch(error => res.status(400).json({ error }));
+            }
         })
         .catch(error => res.status(500).json({ error}));
 };
